Разделил лапшикод на более мелкие файлы
Переместил тесты в зависимости от того какой метод контроллера они тестируют
