+++ tests/Feature/DevTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature\NewTest;
+namespace Tests\Feature;
 
 use App\Models\Profile;
 use App\Models\User;
