+++ tests/Feature/RoutTests/PutPatchTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature\NewTest;
+namespace Tests\Feature\RoutTests;
 
 use App\Models\Profile;
 use App\Models\Skill;
